+++ WoloxAndroidBootstrap/app/src/main/java/ar/com/wolox/android/example/utils/ActivityExtensionFunctions.kt
@@ -1,0 +1,8 @@
+package ar.com.wolox.android.example.utils
+
+import android.app.Activity
+import android.view.View
+
+fun Activity.hideKeyboard() {
+    hideKeyboard(currentFocus ?: View(this))
+}
