+++ WoloxAndroidBootstrap/app/src/main/java/ar/com/wolox/android/example/utils/ContextExtensionFunctions.kt
@@ -1,0 +1,11 @@
+package ar.com.wolox.android.example.utils
+
+import android.app.Activity
+import android.content.Context
+import android.view.View
+import android.view.inputmethod.InputMethodManager
+
+fun Context.hideKeyboard(view: View) {
+    val inputMethodManager = getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
+    inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
+}
