--- WoloxAndroidBootstrap/app/src/main/java/ar/com/wolox/android/example/utils/ExtensionFunctions.kt
@@ -1,27 +1,0 @@
-package ar.com.wolox.android.example.utils
-
-import android.app.Activity
-import android.content.Context
-import android.text.TextUtils
-import android.util.Patterns
-import android.view.View
-import android.view.inputmethod.InputMethodManager
-import androidx.fragment.app.Fragment
-import com.google.android.material.snackbar.Snackbar
-
-fun String.isEmailValid(): Boolean {
-    return !TextUtils.isEmpty(this) && Patterns.EMAIL_ADDRESS.matcher(this).matches()
-}
-
-fun Fragment.showSnackbar(message: String, length: Int = Snackbar.LENGTH_SHORT) = view?.run { Snackbar.make(this, message, length).show() }
-
-fun Fragment.hideKeyboard() {
-    view?.let { activity?.hideKeyboard(it) }
-}
-fun Activity.hideKeyboard() {
-    hideKeyboard(currentFocus ?: View(this))
-}
-fun Context.hideKeyboard(view: View) {
-    val inputMethodManager = getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
-    inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
-}
